+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-C7JYVbgC.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-D77HNypK.js";import{_ as a}from"./mermaid.core-DF5fAwWL.js";import"./chunk-5VM5RSS4-ULOrlwHF.js";import"./chunk-XXDRQBXY-DGw1iCXl.js";import"./chunk-POPQ4Y6H-GuaOu8cs.js";import"./chunk-F27PBJKO-CLZ-zcQB.js";import"./channel-BokhyPQZ.js";import"./index-SgnWGC6o.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};